Refac: move boatMakerTypes to helpers
