+++ src/components/maps/leafletHelpers.js
@@ -1,4 +1,5 @@
 import * as L from 'leaflet'
+import { appWindAngle } from '../../utils/angleFormatter.js'
 
 // Basic maps providers
 export const baseMaps = function () {
@@ -80,4 +81,81 @@ export const overlayMaps = function () {
   }
 
   return overlayMaps
+}
+
+export const boatMarkerTypes = function () {
+  const sailConfigImage = (twd, cog) => {
+    const [awa, tack] = appWindAngle(twd, cog)
+    if (awa > 155) {
+      return `/sailboat-180${tack}.png` // running
+    } else if (awa > 110) {
+      return `/sailboat-135${tack}.png` // broad reaching
+    } else if (awa > 50) {
+      return `/sailboat-090${tack}.png` // close/beam reaching
+    } else if (awa > 25) {
+      return `/sailboat-045${tack}.png` // close-hauled, beating
+    } else {
+      return '/sailboat-000.png'
+    }
+  }
+  const sailConfigIconImg = function (feature) {
+    if (
+      feature.properties.status == 'sailing' &&
+      feature.properties.truewinddirection &&
+      feature.properties.courseovergroundtrue
+    ) {
+      return sailConfigImage(feature.properties.truewinddirection, feature.properties.courseovergroundtrue)
+    }
+    return '/sailboat-000.png'
+  }
+
+  const sailBoatConfigIcon = function (feature, latlng) {
+    return L.marker(latlng, {
+      icon: new L.Icon({
+        iconSize: [32, 32],
+        iconAnchor: [16, 10],
+        iconUrl: sailConfigIconImg(feature),
+      }),
+      rotationAngle: feature.properties.courseovergroundtrue,
+    })
+  }
+  const sailBoatIcon = function (feature, latlng) {
+    return L.marker(latlng, {
+      icon: new L.Icon({
+        iconSize: [16, 32],
+        iconAnchor: [8, 10],
+        iconUrl: '/sailboaticon.png',
+      }),
+      rotationAngle: feature.properties.courseovergroundtrue,
+    })
+  }
+  const powerBoatIcon = function (feature, latlng) {
+    return L.marker(latlng, {
+      icon: new L.Icon({
+        iconSize: [16, 32],
+        iconAnchor: [8, 10],
+        iconUrl: '/powerboaticon.png',
+      }),
+      rotationAngle: feature.properties.courseovergroundtrue,
+    })
+  }
+  const simpleDotIcon = function (feature, latlng) {
+    return L.circleMarker(latlng, {
+      radius: 2,
+      fillColor: '#00FFFF',
+      color: '#000',
+      weight: 1,
+      opacity: 1,
+      fillOpacity: 0.8,
+    })
+  }
+
+  const boatTypes = {
+    Sailboat: sailBoatIcon,
+    SailboatConfig: sailBoatConfigIcon,
+    Powerboat: powerBoatIcon,
+    Dot: simpleDotIcon,
+  }
+
+  return boatTypes
 }
