+++ src/utils/angleFormatter.js
@@ -7,21 +7,6 @@ export const appWindAngle = (twd, cog) => {
   return [awa, tack]
 }
 
-export const sailConfigImage = (twd, cog) => {
-  const [awa, tack] = appWindAngle(twd, cog)
-  if (awa > 155) {
-    return `/sailboat-180${tack}.png` // running
-  } else if (awa > 110) {
-    return `/sailboat-135${tack}.png` // broad reaching
-  } else if (awa > 50) {
-    return `/sailboat-090${tack}.png` // close/beam reaching
-  } else if (awa > 25) {
-    return `/sailboat-045${tack}.png` // close-hauled, beating
-  } else {
-    return '/sailboat-000.png'
-  }
-}
-
 export const angleFormat = (val) => {
   if (!val) return null
   const deg = parseFloat(val)
